Refactor Client as interface, ClientImpl as implementation.
